refactor(app): share operator assets checks
