+++ openspec/changes/refactor-app-operator-assets/README.md
@@ -1,0 +1,3 @@
+# refactor-app-operator-assets
+
+Refactor: share operator assets checks between CLI and MCP
