+++ src/app/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod next_actions;
+pub(crate) mod operator_assets;
 pub(crate) mod preview_diff;
